[FEATURE] Streamline default expiration date
The bash script sets the expiration date to "none"
by default. The web form has "month" as default however.

Streamline the default expiration date and set the bash
to month as well.

This also avoids an error if the bash script is run
for the first time ever with default settings, since
the expiration time "none" is not allowed then.

Refs #85