[FEATURE] Bash script: Enhance help and add version
Enhance the help and show all available options in
a common structure.
Move description of available options to one place.

Also add the version of the Jirafeau version which
generated the bash script. This way a user may
find out about changes (because his script was
generated with Jirafeau 1.1.2, but the current version
on the server may be 3.2.5 already).