[BUGFIX] Bash script: return download link to web view
The bash script uses the upload URL to build the download link.

So the user may share different links when using the bash script
or the webform. Besides that does the API page show the same
result for "download links" and "direct links", so the current
list makes no sense to the uploader.

The script should however link to the web view and
offer API commands additionally.

Refs #86