[FEATURE] Install script: Reduce test code
Add some comments to enhance the readabillity in the
installation script.

Reduce footer includes and send all fatal errors
to one function instead.

Remove includes of installation depenend variables
in the footer, if the installation script is not
done yet.

Move settings include to the top (this may help
for future configuration tests).