[!!][FEATURE] Let admin overwrite the ToS page
Let the admin overwrite the content of the
"Terms of Service" page by copying the original ToS
template and creating a local file (same procedure as
for the configuration). This file is not versionized.

Note: This is a breaking change, since the existing
tos textfile is moved. This should be mentioned
in the update comments.
  * ToS text file changed → move file from "/tos_text.php"
    to "/lib/tos.local.txt" and remove all HTML und PHP Tags,
    leaving a regular text file.

Refs #95