[!!][FEATURE] Remove XHTML doctype
The project uses a HTML5 doctype, but switches
to XHTML if a condition says that the clients browser
can handle XHTML. As this is the case for nearly any
browser which also supports HTML5, the condition causes
the script to *always* render the XHTML doctype.

We remove the whole condition and the XHTML doctype
in favour of a HTML5 doctype only.

Note: This may be considered as breaking change,
since HTML5 may not be working in ancient versions of IE.

Refs #97