[FEATURE] Show readable date format
Whenever date objects in PHP or JavaScript are converted
to strings, it depends on the servers/clients settings and
timezone which locale and format is used.

This may lead to situation where the date is in english,
but all other texts of Jirafeau service are in another language.

Depending on unser language it may be not possible to localise
the date, since we would need to have a list of all existing
date formats. In both PHP and JavaScript.

To avoid these translation obstacles we may use the following
fallback:
  * Always use UTC timezone in PHP and render dates with UTC only
  * Use the rather international format »YYYY-MM-DD hh:mm«

Refs #59 #106 #73